# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LLM Router native API
  version: 1.0.0
  summary: Provider-neutral routing, agent, tool, accounting, and administration API.
  description: >-
    This is the complete first-release HTTP JSON contract. Stream event details
    are in stream-protocol.md. Contract envelopes are closed. A declared JSON
    document field follows its own supplied schema.
servers:
  - url: /
security:
  - bearerToken: []
...
paths:
...
  /v1/admin/session-starts:
    post:
      tags: [Authentication]
      operationId: startAdministratorSession
      summary: Start a Pocket ID login or recent-authentication flow.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdministratorSessionStart'}
      responses:
        '201':
          description: Authorization redirect created with server-held state, nonce, and PKCE verifier.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdministratorAuthorization'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '503': {$ref: '#/components/responses/Retryable'}
  /v1/admin/oidc/callback:
    get:
      tags: [Authentication]
      operationId: completeAdministratorSession
      summary: Validate the Pocket ID callback and create or refresh the local session.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 4096}
        - name: state
          in: query
          required: true
          schema: {type: string, minLength: 32, maxLength: 4096}
      responses:
        '303':
          description: Authentication completed. The response sets the local session cookie and redirects to the validated local return path.
          headers:
            Location:
              required: true
              schema: {type: string}
            Set-Cookie:
              required: true
              schema: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/Retryable'}
  /v1/admin/session:
    get:
      tags: [Authentication]
      operationId: getAdministratorSession
      summary: Get the local administrator session and current local grants.
      security: [{administratorSession: []}]
      responses:
        '200':
          description: Current local session without an identity-provider token.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdministratorSession'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    delete:
      tags: [Authentication]
      operationId: deleteAdministratorSession
      summary: Revoke the local administrator session and clear its cookie.
      security: [{administratorSession: []}]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/Origin'
      responses:
        '204': {description: Local session revoked.}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
    administratorSession:
      type: apiKey
      in: cookie
      name: __Host-llmrouter-admin
    embedSession:
      type: apiKey
      in: cookie
      name: __Host-llmrouter-embed
      description: >-
        Write-only browser authority set during bootstrap. It is valid only
        from the exact Router frame origin bound to the unexpired,
        non-revoked embed session and cannot authorize global administration.
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 32, maxLength: 500}
    Origin:
      name: Origin
      in: header
      required: true
      schema: {type: string, format: uri, pattern: '^https?://[^/]+$'}
...
  responses:
...
    BadRequest:
      description: Invalid request.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    Unauthorized:
      description: Authentication failed or is too old.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    Forbidden:
      description: The authenticated actor lacks exact scope.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
...
    Retryable:
      description: Temporary failure. Retry only as errors.md permits.
      headers:
        Retry-After:
          schema: {type: integer, minimum: 1}
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
...
  schemas:
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 200
...
    Timestamp:
      type: string
      format: date-time
...
    AdministratorSessionStart:
      type: object
      additionalProperties: false
      required: [purpose, return_path]
      properties:
        purpose: {type: string, enum: [login, recent_authentication]}
        return_path:
          type: string
          pattern: '^/[A-Za-z0-9._~!$&''()*+,;=:@%/?-]*$'
          maxLength: 2000
        trusted_grant_token:
          type: string
          minLength: 43
          maxLength: 43
          pattern: '^[A-Za-z0-9_-]+$'
          description: One-use trusted-console grant token for initial or recovery login.
    AdministratorAuthorization:
      type: object
      additionalProperties: false
      required: [authorization_url, expires_at]
      properties:
        authorization_url: {type: string, format: uri}
        expires_at: {$ref: '#/components/schemas/Timestamp'}
    AdministratorSession:
      type: object
      additionalProperties: false
      required: [issuer, subject, grants, authenticated_at, recent_authentication_at, account_state_checked_at, idle_expires_at, absolute_expires_at, csrf_token, identity_account_url, authentication_mode]
      properties:
        issuer: {type: string, format: uri}
        subject: {$ref: '#/components/schemas/OpaqueId'}
        grants:
          type: array
          uniqueItems: true
          items: {type: string, maxLength: 200}
        authenticated_at: {$ref: '#/components/schemas/Timestamp'}
        account_state_checked_at: {$ref: '#/components/schemas/Timestamp'}
        recent_authentication_at:
          oneOf: [{$ref: '#/components/schemas/Timestamp'}, {type: 'null'}]
        idle_expires_at: {$ref: '#/components/schemas/Timestamp'}
        absolute_expires_at: {$ref: '#/components/schemas/Timestamp'}
        csrf_token: {type: string, minLength: 32, maxLength: 200, pattern: '^[A-Za-z0-9_-]+$'}
        identity_account_url: {type: string, format: uri}
        authentication_mode: {type: string, enum: [local, oidc]}
...
    FieldError:
      type: object
      additionalProperties: false
      required: [path, code, message]
      properties:
        path: {type: string, maxLength: 500}
        code: {type: string, maxLength: 100}
        message: {type: string, maxLength: 1000}
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable, request_id]
          properties:
            code:
              type: string
              enum: [invalid_request, unsupported_contract, unsupported_capability, invalid_token, recent_auth_required, insufficient_scope, service_scope_mismatch, workspace_scope_mismatch, policy_denied, diagnostic_permission_required, not_found, request_not_found, workspace_not_found, attachment_not_found, idempotency_conflict, state_revision_conflict, configuration_revision_conflict, request_identity_conflict, stream_replay_unavailable, terminal_state, budget_ceiling_conflict, attachment_already_complete, request_identity_expired, workspace_retired, assignment_unavailable, workspace_unavailable, capability_mismatch, embedding_space_mismatch, budget_exhausted, secret_detected, attachment_invalid, rate_limited, temporarily_unavailable, stale_configuration, spool_capacity_exhausted, allowance_unavailable, internal_error]
            message: {type: string, maxLength: 1000}
            retryable: {type: boolean}
            request_id: {$ref: '#/components/schemas/OpaqueId'}
            retry_after_ms: {type: [integer, 'null'], minimum: 0}
            field_errors: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/FieldError'}}